Update circle.py
Added attributes x0, y0 (co-ordinates of centre of circle) to Circle Class object instance. A Circle object is defined by the centre (x0, y0) in the constructor. However, there were no attributes corresponding to this only x, y which are x0-r, y0-r due to inheritance from RoundRect (most shapes are defined by the top-left corner co-ordinate x, y). The addition of the centre of Circle attributes make it easier to locate the circle on the screen using the same reference co-ordinates used to define it.